requirements.txt
# Backend Dependencies
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
requests

# Frontend Dependencies
streamlit
pandas
plotly